Adds a brief overview to the README.md
